+++ backend/ws-server/src/main/java/org/linkwave/ws/repository/ChatRepository.java
@@ -1,4 +1,4 @@
-package org.linkwave.ws.websocket.repository;
+package org.linkwave.ws.repository;
 
 import org.linkwave.ws.api.chat.ChatMemberDto;
