+++ backend/ws-server/src/main/java/org/linkwave/ws/websocket/route/GroupChatRoutes.java
@@ -10,7 +10,7 @@ import org.linkwave.ws.websocket.dto.ChatMessage;
 import org.linkwave.ws.websocket.dto.ErrorMessage;
 import org.linkwave.ws.websocket.dto.NewGroupChat;
 import org.linkwave.ws.websocket.jwt.UserPrincipal;
-import org.linkwave.ws.websocket.repository.ChatRepository;
+import org.linkwave.ws.repository.ChatRepository;
 import org.linkwave.ws.websocket.routing.Box;
 import org.linkwave.ws.websocket.routing.Payload;
 import org.linkwave.ws.websocket.routing.bpp.Broadcast;
