+++ backend/ws-server/src/main/java/org/linkwave/ws/websocket/routing/broadcast/instances/MessageDelegateImpl.java
@@ -4,7 +4,7 @@ import org.linkwave.ws.websocket.dto.Action;
 import org.linkwave.ws.websocket.dto.BaseMessage;
 import org.linkwave.ws.websocket.dto.OutcomeMessage;
 import org.linkwave.ws.websocket.dto.StatusMessage;
-import org.linkwave.ws.websocket.repository.ChatRepository;
+import org.linkwave.ws.repository.ChatRepository;
 import org.linkwave.ws.websocket.routing.broadcast.WebSocketMessageBroadcast;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
