+++ backend/ws-server/src/test/java/org/linkwave/ws/unit/GroupChatRoutesTest.java
@@ -11,7 +11,7 @@ import org.linkwave.ws.websocket.dto.ChatMessage;
 import org.linkwave.ws.websocket.dto.ErrorMessage;
 import org.linkwave.ws.websocket.dto.NewGroupChat;
 import org.linkwave.ws.websocket.jwt.UserPrincipal;
-import org.linkwave.ws.websocket.repository.ChatRepository;
+import org.linkwave.ws.repository.ChatRepository;
 import org.linkwave.ws.websocket.route.GroupChatRoutes;
 import org.linkwave.ws.websocket.routing.Box;
 import org.mockito.InjectMocks;
