+++ backend/ws-server/src/test/java/org/linkwave/ws/unit/SimpleBroadcastManagerTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.linkwave.ws.websocket.repository.ChatRepository;
+import org.linkwave.ws.repository.ChatRepository;
 import org.linkwave.ws.websocket.routing.bpp.Broadcast;
 import org.linkwave.ws.websocket.routing.broadcast.SimpleBroadcastManager;
 import org.linkwave.ws.websocket.routing.broadcast.WebSocketMessageBroadcast;
